+++ CHANGELOG.md
@@ -9,6 +9,20 @@ See [GitHub releases](https://github.com/mll-lab/laravel-utils/releases).
 
 ## Unreleased
 
+## v4.0.0
+
+### Added
+
+- Add conditional migrations
+
+### Removed
+
+- Drop support for Laravel 8
+
+### Changed
+
+- Use union and intersection types in `ModelStates`
+
 ## v3.0.0
 
 ### Changed